api/requirements.txt
fastapi
uvicorn
FastAPI-SQLAlchemy
psycopg2-binary
